+++ src/Controller/Admin/MenuCrudController.php
@@ -3,6 +3,7 @@
 namespace App\Controller\Admin;
 
 use App\Entity\Menu;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Actions;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
@@ -11,6 +12,8 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 #[IsGranted('ROLE_ADMIN')]
 class MenuCrudController extends AbstractCrudController
 {
+    use AdminCrudPolishTrait;
+
     public static function getEntityFqcn(): string
     {
         return Menu::class;
@@ -18,9 +21,14 @@ class MenuCrudController extends AbstractCrudController
 
     public function configureCrud(Crud $crud): Crud
     {
-        return $crud
+        return $this->inlineRowActions($crud
             ->setEntityLabelInSingular('admin.menu_entity.entity.singular')
-            ->setEntityLabelInPlural('admin.menu_entity.entity.plural');
+            ->setEntityLabelInPlural('admin.menu_entity.entity.plural'));
+    }
+
+    public function configureActions(Actions $actions): Actions
+    {
+        return $this->addBackToListAction($actions);
     }
 
     public function configureFields(string $pageName): iterable