+++ src/Controller/Admin/MenuItemCrudController.php
@@ -3,6 +3,7 @@
 namespace App\Controller\Admin;
 
 use App\Entity\MenuItem;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Actions;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
@@ -13,6 +14,8 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 #[IsGranted('ROLE_ADMIN')]
 class MenuItemCrudController extends AbstractCrudController
 {
+    use AdminCrudPolishTrait;
+
     public static function getEntityFqcn(): string
     {
         return MenuItem::class;
@@ -20,10 +23,15 @@ class MenuItemCrudController extends AbstractCrudController
 
     public function configureCrud(Crud $crud): Crud
     {
-        return $crud
+        return $this->inlineRowActions($crud
             ->setEntityLabelInSingular('admin.menu_item.entity.singular')
             ->setEntityLabelInPlural('admin.menu_item.entity.plural')
-            ->setDefaultSort(['position' => 'ASC']);
+            ->setDefaultSort(['position' => 'ASC']));
+    }
+
+    public function configureActions(Actions $actions): Actions
+    {
+        return $this->addBackToListAction($actions);
     }
 
     public function configureFields(string $pageName): iterable